get all clients

+++ API_wee/Models/Policy.cs
@@ -7,9 +7,9 @@ namespace API_wee.Models
     {
         [Key]
         public int Id_policy { get; set; }
-        public short Id_client { get; set; }
-        public short Id_typePolicy { get; set; }
-        public short id_statusPolicy { get; set; }
+        public int Id_client { get; set; }
+        public int Id_typePolicy { get; set; }
+        public int id_statusPolicy { get; set; }
         public string numPolicy { get; set; } = string.Empty;
         public DateTime startDatePolicy { get; set; } 
         public DateTime endDatePolicy { get; set; }
